# phantom_mask_laradock/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phantom Mask API",
        "description": "藥局口罩平台 API 文件",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://phantom_mask_laradock.kdan.succ.work",
            "description": "Development API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://phantom_mask_laradock.kdan.succ.work",
            "description": "Production API Server"
        }
    ],
    "paths": {
        "/api/pharmacies": {
            "get": {
                "tags": [
                    "藥局"
                ],
                "summary": "取得藥局列表，支援搜尋、過濾和排序",
                "operationId": "5888c7633db75f5dd2995d7866ba44ab",
                "parameters": [
                    {
                        "name": "day",
                        "in": "query",
                        "description": "星期幾 (Monday/Tuesday/Wednesday/Thursday/Friday/Saturday/Sunday)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Monday",
                                "Tuesday",
                                "Wednesday",
                                "Thursday",
                                "Friday",
                                "Saturday",
                                "Sunday"
                            ]
                        }
                    },
                    {
                        "name": "time",
                        "in": "query",
                        "description": "時間 (HH:mm)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "time",
                            "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "搜尋關鍵字（藥局名稱或口罩名稱）",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "最低價格",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "最高價格",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "mask_count",
                        "in": "query",
                        "description": "最少口罩數量",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功取得藥局列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "DFW Wellness"
                                                    },
                                                    "cash_balance": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 328.41
                                                    },
                                                    "opening_hours": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "day": {
                                                                    "type": "string",
                                                                    "example": "Monday"
                                                                },
                                                                "open": {
                                                                    "type": "string",
                                                                    "example": "08:00"
                                                                },
                                                                "close": {
                                                                    "type": "string",
                                                                    "example": "17:00"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "masks": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Black Mask"
                                                                },
                                                                "price": {
                                                                    "type": "number",
                                                                    "format": "float",
                                                                    "example": 10
                                                                },
                                                                "stock": {
                                                                    "description": "庫存數量",
                                                                    "type": "integer",
                                                                    "example": 100
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證錯誤",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "驗證失敗"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "查詢藥局失敗"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}